Move reading of data from text pages into "PageDataExtractor"
* Takes the list of text pages and returns a list of the data elements (person, status, etc.)
